+++ resources/views/consultations/show.blade.php
@@ -14,10 +14,12 @@
         <p>Profesional responsable: {{ $consultation->user->name }}</p>
     </div>
 
-    {!! Form::open(['method' => 'DELETE', 'route' => ['paciente.consultas.destroy', $patient->id, $consultation->id ]]) !!}
-        <div class="form-group btn-group" role="group">
-            {!! link_to_route('paciente.consultas.edit', 'Editar', [$patient->id, $consultation->id], ['class' => 'btn btn-primary']) !!}
-            {!! Form::button('<i class="fa fa-trash-o fa-fw"></i>&nbsp;Borrar', ['class' => 'btn btn-danger', 'type' => 'submit']) !!}
-        </div>
-    {!! Form::close() !!}
+    @can('update_consultation', $consultation)
+        {!! Form::open(['method' => 'DELETE', 'route' => ['paciente.consultas.destroy', $patient->id, $consultation->id ]]) !!}
+            <div class="form-group btn-group" role="group">
+                {!! link_to_route('paciente.consultas.edit', 'Editar', [$patient->id, $consultation->id], ['class' => 'btn btn-primary']) !!}
+                {!! Form::button('<i class="fa fa-trash-o fa-fw"></i>&nbsp;Borrar', ['class' => 'btn btn-danger', 'type' => 'submit']) !!}
+            </div>
+        {!! Form::close() !!}
+    @endcan
 @stop
