+++ resources/views/controls/show.blade.php
@@ -20,10 +20,12 @@
         <p>Profesional responsable: {{ $control->user->name }}</p>
     </div>
 
-    {!! Form::open(['method' => 'DELETE', 'route' => ['paciente.controles.destroy', $patient->id, $control->id ]]) !!}
-        <div class="form-group btn-group" role="group">
-            {!! link_to_route('paciente.controles.edit', 'Editar', [$patient->id, $control->id], ['class' => 'btn btn-primary']) !!}
-            {!! Form::button('<i class="fa fa-trash-o fa-fw"></i>&nbsp;Delete', ['class' => 'btn btn-danger', 'type' => 'submit']) !!}
-        </div>
-    {!! Form::close() !!}
+    @can('update_control', $control)
+        {!! Form::open(['method' => 'DELETE', 'route' => ['paciente.controles.destroy', $patient->id, $control->id ]]) !!}
+            <div class="form-group btn-group" role="group">
+                {!! link_to_route('paciente.controles.edit', 'Editar', [$patient->id, $control->id], ['class' => 'btn btn-primary']) !!}
+                {!! Form::button('<i class="fa fa-trash-o fa-fw"></i>&nbsp;Delete', ['class' => 'btn btn-danger', 'type' => 'submit']) !!}
+            </div>
+        {!! Form::close() !!}
+    @endcan
 @stop
