+++ resources/views/patients/partials/_patientDetails.blade.php
@@ -32,10 +32,12 @@
         </div>
     </form>
 
-    {!! Form::open(['method' => 'DELETE', 'route' => ['paciente.destroy', $patient->id]]) !!}
-    <div class="form-group btn-group" role="group">
-        {!! Form::button('<i class="fa fa-trash-o fa-fw"></i>&nbsp;Borrar', ['class' => 'btn btn-danger', 'type' => 'submit']) !!}
-        {!! link_to_route('paciente.edit', 'Editar', [$patient->id], ['class' => 'btn btn-primary']) !!}
-    </div>
-    {!! Form::close() !!}
+    @can('update_patient', $patient)
+        {!! Form::open(['method' => 'DELETE', 'route' => ['paciente.destroy', $patient->id]]) !!}
+            <div class="form-group btn-group" role="group">
+                {!! Form::button('<i class="fa fa-trash-o fa-fw"></i>&nbsp;Borrar', ['class' => 'btn btn-danger', 'type' => 'submit']) !!}
+                {!! link_to_route('paciente.edit', 'Editar', [$patient->id], ['class' => 'btn btn-primary']) !!}
+            </div>
+        {!! Form::close() !!}
+    @endcan
 </div>
